# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce Demo API",
    "version": "1.0.0",
    "description": "花卉電商網站 REST API"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Id"
      }
    }
  },
  "paths": {
    "/api/admin/orders": {
      "get": {
        "summary": "後台訂單列表",
        "tags": [
          "Admin Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "paid",
                "failed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "orders": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "order_no": {
                                "type": "string"
                              },
                              "user_id": {
                                "type": "string"
                              },
                              "recipient_name": {
                                "type": "string"
                              },
                              "recipient_email": {
                                "type": "string"
                              },
                              "total_amount": {
                                "type": "integer"
                              },
                              "status": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "integer"
                            },
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/orders/{id}": {
      "get": {
        "summary": "後台訂單詳情",
        "tags": [
          "Admin Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "order_no": {
                          "type": "string"
                        },
                        "user_id": {
                          "type": "string"
                        },
                        "recipient_name": {
                          "type": "string"
                        },
                        "recipient_email": {
                          "type": "string"
                        },
                        "recipient_address": {
                          "type": "string"
                        },
                        "total_amount": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "product_id": {
                                "type": "string"
                              },
                              "product_name": {
                                "type": "string"
                              },
                              "product_price": {
                                "type": "integer"
                              },
                              "quantity": {
                                "type": "integer"
                              }
                            }
                          }
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "訂單不存在"
          }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "summary": "從購物車建立訂單",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recipientName",
                  "recipientEmail",
                  "recipientAddress"
                ],
                "properties": {
                  "recipientName": {
                    "type": "string"
                  },
                  "recipientEmail": {
                    "type": "string",
                    "format": "email"
                  },
                  "recipientAddress": {
                    "type": "string"
                  },
                  "shippingMethod": {
                    "type": "string",
                    "enum": [
                      "home_delivery",
                      "cvs"
                    ],
                    "description": "配送方式，預設 home_delivery（宅配）。宅配基本運費 120 元，商品小計滿 1,500 元免收；超商取貨（cvs）固定 60 元，不受滿額免運影響。"
                  },
                  "isRemote": {
                    "type": "boolean",
                    "description": "是否為偏遠地區，加收 200 元，預設 false"
                  },
                  "isUrgent": {
                    "type": "boolean",
                    "description": "是否為當日急件，加收 250 元，預設 false"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "訂單建立成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "order_no": {
                          "type": "string"
                        },
                        "product_subtotal": {
                          "type": "integer"
                        },
                        "shipping_fee": {
                          "type": "integer"
                        },
                        "shipping_method": {
                          "type": "string"
                        },
                        "is_remote": {
                          "type": "boolean"
                        },
                        "is_urgent": {
                          "type": "boolean"
                        },
                        "total_amount": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "product_name": {
                                "type": "string"
                              },
                              "product_price": {
                                "type": "integer"
                              },
                              "quantity": {
                                "type": "integer"
                              }
                            }
                          }
                        },
                        "created_at": {
                          "type": "string"
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "購物車為空或庫存不足或收件資訊缺失"
          }
        }
      },
      "get": {
        "summary": "自己的訂單列表",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "orders": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "order_no": {
                                "type": "string"
                              },
                              "total_amount": {
                                "type": "integer"
                              },
                              "status": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/{id}/check-payment": {
      "post": {
        "summary": "透過綠界 QueryTradeInfo API 查詢付款狀態",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "訂單狀態不是 pending"
          },
          "404": {
            "description": "訂單不存在"
          }
        }
      }
    },
...
  },
...
}